+++ src/App.tsx
@@ -11,19 +11,18 @@ import {
   LectureDetail,
 } from './pages'
 import Yoyo from './pages/Yoyo'
-import YoyoComment from './pages/Yoyo/YoyoComment'
 import { LayoutRoute } from './components'
-import YoyoReview from './pages/Yoyo/YoyoReview'
-
 import eiei from './pages/Yoyo/user'
 import Post from './pages/Yoyo/Post'
+
 const App: React.FC = () => {
   return (
     <>
       <BrowserRouter>
         <Switch>
           <Route exact path="/">
-            <Redirect to="/login" />
+            {/* ชั่วคราวสำหรับ test */}
+            <Redirect to="/yoyo" />
           </Route>
           <Route exact path="/login" component={Login} />
           <Route exact path="/register" component={Register} />
@@ -34,9 +33,6 @@ const App: React.FC = () => {
           <LayoutRoute exact path="/profile" component={Profile} />
           {/* for test */}
           <LayoutRoute exact path="/yoyo" component={Yoyo} />
-          {/* <LayoutRoute path="/yoyoComment" component={YoyoComment} /> */}
-          <LayoutRoute exact path="/yoyoComment/:id" component={YoyoComment} />
-          <LayoutRoute exact path="/yoyoReview" component={YoyoReview} />
           <LayoutRoute exact path="/post/:id" component={Post} />
           <LayoutRoute exact path="/pong" component={eiei} />
           <Route exact path="*" component={NotFound} />
